+++ docs/components/Header.js
@@ -1,6 +1,8 @@
 // @flow strict
 import { type Node, useCallback, useEffect, useState } from 'react';
 import { Box, Flex, FixedZIndex, Text, IconButton, Sticky } from 'gestalt';
+// $FlowExpectedError[untyped-import]
+import GestaltPackageJson from 'gestalt/package.json';
 import DocSearch from './DocSearch.js';
 import GestaltLogo from './GestaltLogo.js';
 import HeaderMenu from './HeaderMenu.js';
@@ -22,7 +24,7 @@ function Header() {
       alignItems="center"
       role="banner"
     >
-      <Box marginStart={-2} marginEnd={-2}>
+      <Box marginStart={-2} marginEnd={2} display="flex" alignItems="center">
         {/* <Text> is out here to get proper underline styles on link */}
         <Text color="darkGray" weight="bold">
           <Link
@@ -34,10 +36,15 @@ function Header() {
               <Flex alignItems="center" gap={2}>
                 <GestaltLogo height={40} width={40} />
                 {/* slight tweak to vertically center to logo */}
-                <Box dangerouslySetInlineStyle={{ __style: { marginBottom: '1px' } }}>Gestalt</Box>
+                <Box dangerouslySetInlineStyle={{ __style: { marginBottom: '1px' } }}>
+                  Gestalt Design System
+                </Box>
               </Flex>
             </Box>
           </Link>
+        </Text>
+        <Text size="sm" weight="bold">
+          v{GestaltPackageJson.version}
         </Text>
       </Box>