+++ packages/gestalt/src/behaviors/TrapFocusBehavior.js
@@ -1,11 +1,12 @@
 // @flow strict
-import { type Node as ReactNode, Component } from 'react';
+import { useEffect, useRef, type Node as ReactNode } from 'react';
 
 type Props = {|
   children?: ReactNode,
 |};
 
 function queryFocusableAll(el: HTMLDivElement) {
+  // Focusable, interactive elements that could possibly be in children
   const selector = [
     'a[href]',
     'area[href]',
@@ -27,54 +28,69 @@ function queryFocusableAll(el: HTMLDivElement) {
 }
 
 const focusElement = (el: HTMLElement) => {
+  // https://github.com/facebook/flow/issues/8705
   // $FlowFixMe[method-unbinding]
   if (typeof el.focus === 'function') {
     el.focus();
   }
 };
 
-export default class TrapFocusBehavior extends Component<Props> {
-  el: ?HTMLDivElement;
+/**
+ * TrapFocusBehavior is used by components like Modal and Sheet to ensure that only elements within children components can be focused.
+ */
+export default function TrapFocusBehavior({ children }: Props): ReactNode {
+  const elRef = useRef<?HTMLDivElement>(null);
+  const previouslyFocusedElRef = useRef<?HTMLElement>(null);
 
-  previouslyFocusedEl: ?HTMLElement;
-
-  componentDidMount() {
-    this.previouslyFocusedEl = document.activeElement;
-    this.focusFirstChild();
-    document.addEventListener('focus', this.handleFocus, true);
-  }
-
-  componentWillUnmount() {
-    document.removeEventListener('focus', this.handleFocus, true);
-    if (this.previouslyFocusedEl) {
-      focusElement(this.previouslyFocusedEl);
-    }
-  }
-
-  setElRef: (el: ?HTMLDivElement) => void = (el: ?HTMLDivElement) => {
+  const setElRef: (el: ?HTMLDivElement) => void = (el: ?HTMLDivElement) => {
     if (el) {
-      this.el = el;
+      elRef.current = el;
     }
   };
 
-  handleFocus: (event: FocusEvent) => void = (event: FocusEvent) => {
-    if (!this.el || (event.target instanceof Node && this.el.contains(event.target))) {
-      return;
-    }
+  useEffect(() => {
+    const { current: element } = elRef;
 
-    event.stopPropagation();
-    event.preventDefault();
-    this.focusFirstChild();
-  };
+    // Focus the first child element among all the focusable, interactive elements within `children`
+    const focusFirstChild = () => {
+      if (element) {
+        focusElement(queryFocusableAll(element)[0]);
+      }
+    };
 
-  focusFirstChild() {
-    const { el } = this;
-    if (el) {
-      focusElement(queryFocusableAll(el)[0]);
-    }
-  }
+    const handleFocus: (event: FocusEvent) => void = (event: FocusEvent) => {
+      if (!element || (event.target instanceof Node && element.contains(event.target))) {
+        return;
+      }
 
-  render(): ReactNode {
-    return <div ref={this.setElRef}>{this.props.children}</div>;
-  }
+      // This prevents stack overflow when multiple TrapFocusBehaviors are rendered
+      if (event.target instanceof Element && event.target.closest('[name="trap-focus"]') !== null) {
+        return;
+      }
+
+      event.stopPropagation();
+      event.preventDefault();
+      focusFirstChild();
+    };
+
+    // If an element has focus currently, keep a reference to that element
+    previouslyFocusedElRef.current = document.activeElement;
+    focusFirstChild();
+    document.addEventListener('focus', handleFocus, true);
+
+    return function cleanup() {
+      const { current: previouslyFocusedEl } = previouslyFocusedElRef;
+      document.removeEventListener('focus', handleFocus, true);
+      // If we previously stored a reference to a focused element, return focus to that element
+      if (previouslyFocusedEl) {
+        focusElement(previouslyFocusedEl);
+      }
+    };
+  }, [elRef, previouslyFocusedElRef]);
+
+  return (
+    <div name="trap-focus" ref={setElRef}>
+      {children}
+    </div>
+  );
 }